// src/components/faq/types.ts
/**
 * This code was generated by Builder.io.
 * https://builder.io/content/undefined?codeSession=78c6445a&sid=c36d2297
 */
export interface FAQItemProps {
  question: string;
  answer?: string;
  isOpen?: boolean;
  onToggle?: () => void;
}

export interface FAQData {
  question: string;
  answer: string;
}